Use a git revision of hd44780-driver which has set_cursor_xy() builtin.
This removes the need for the temporary hd44780_helpers crate.  Note
that 20x4 is the default, so no need to specify dimensions anymore.

The main program keeps using its own helper, as that is published on
crates.io which forbids dependencies on external sources. So we have
to wait for a new release for it to use set_cursor_xy() too.